Update Coupon.php
I found a bug in the delete() method of the Coupon class.

PHP
public function delete(string $coupon_code)
{
    $this->coupon = $this->coupon->find($coupon_code);
    $this->coupon->delete($coupon_code);
}

The problem is that the find() method returns a Coupon object, but the delete() method expects a string. So, the code is trying to delete a Coupon object, which is not possible.

To fix this bug, we need to change the delete() method to accept a Coupon object as its argument. The following code shows the corrected version of the delete() method:

PHP
public function delete(Coupon $coupon)
{
    $coupon->delete();
}

This code first checks if the $coupon argument is an instance of the Coupon class. If it is, then the code calls the delete() method on the $coupon object. Otherwise, the code throws an exception.